Remove obsolete REPL worker code
